Moved invoice_automation inside of huxley, and moved intuit constants to django settings
