+++ huxley/settings/main.py
@@ -146,3 +146,9 @@ SMARTWAIVER_API_KEY = None
 # Celery Configuration Options
 CELERY_RESULT_BACKEND = 'django-db'
 
+# Intuit Keys
+CLIENT_ID = "AB6M8vUqtDAsPaooTFWYk81NztxLFyRQR2ms6B8IL3hkwBttp6"
+CLIENT_SECRET = "wN2KDQA9X7fWTPpkOnO3U87g2dlZ7HRsxnRtv4Iq"
+REFRESH_TOKEN = ""
+REALM_ID = "1404322785"
+ACCESS_TOKEN = ""
